Added new features
Added new features to the project. The user can now add a new project and
add tasks to the project. Added due date div, that checks the current date
and compares it to the due date of the task. If the task is due today,
the background color will change to green. If the task is due in the
future, the background color will change to green.