Added check functions for Today and Upcoming tasks
Added check functions for Today and Upcoming tasks, which check due date.
If the due date is today, the task is added to Today tasks and color is
changed to red. If the due date is upcoming, the task is added to Upcoming
tasks and color is changed to marine.